// src/api/schemas/scheduleListResponseDto.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * Gourmet7 Agent API
 * AI Agent for DeFi operations
 * OpenAPI spec version: 1.0
 */
import type { ScheduleListResponseDtoMeta } from './scheduleListResponseDtoMeta';
import type { ScheduleResponseDto } from './scheduleResponseDto';

export interface ScheduleListResponseDto {
  data: ScheduleResponseDto[];
  meta: ScheduleListResponseDtoMeta;
}